change the level of right click menu and enable click close the botlist, but there are some small problem in auto-close
